requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.16.0
